Start of a rewrite of the UI.
This is only phase one.
Where backend stuff gets added that is only needed for transition and
testing.
The actual UI rewrite comes later this why this beta release is being
used at the moment.

Changelog:
- Added: Translation support for Keys & Comments of a Config Entry
- Added: Compound/Array Override where the "Edit" button can be
overriden in the array. (Only for arrays, compounds get with the
rewrite)
- Changed: Custom Data Types now expect a Datatype class to be present
or we throw a crash instead. (Only affects Carbon config additions)
- Updated: Carbon Config Lib to latest version and fixed bugs
introduced.
- Fixed: issues with string arrays.